IoT Health Monitor - Skripsie 2024
This is the arduino (.ino) file for an IoT Health Monitor and a text file containing all the code used for this project. This was done for my final year project (Skripsie 2024) at the University of Stellenbosch. This code allows measurement of SPO2, Heart Rate and Temperature, and displaying it on a ESP32-C3 web server and external LCD display.